+++ example/lib/panels/helper.dart
@@ -1,6 +1,6 @@
 import 'dart:convert';
 
-import 'package:dio/src/response.dart';
+import 'package:dio/dio.dart';
 import 'package:flutter/material.dart';
 
 import '../dio_helper.dart';